do not try to relax bounds in tightenBounds (when chaning vartype)
- tighten upper bound of a variable that gets a discrete type only if not
  feas-integral or if not integral and current bound is within
  [eps,feastol] above an integral value
- analogeous for lower bounds
- should fix bug 372 without getting bug 283 back

However, an assert may still come up if one changes the type of a variable
to a discrete one and the variable has an upper bound within (eps,feastol)
below an integer value. We now do not change the bound anymore, so that
an assert as in bugreport 283 may come up in var.c at some point.
Analogeous for lower bounds.