fix problem in lpi_cpx with bound changes (bug 453)
Obviously CPLEX does not accept bound changes that are
below 1e-10 (but sets them to 0). 
We add a define that allows to correct an assert checking that
the changed bounds are equal to the given ones (bug 453). 
This is only a fix within lpi_cpx, but SCIP does obviously not
check whether the bounds have been changed to exactly the
value that it has given to the LPI. In any case the change is
below the tolerances, so it should not cause a problem.