save and restore lp status in conflictAnalyzeLP()
conflictAnalyzeLP() now stores the lp status (status, primal
and dual feasible, solution/dual ray values etc.) and
restores it if the lp data was destroyed by the analysis of
the LP. This can happen since conflictAnalyzeLP() solves
modified LPs. By this we should avoid the case in which
the lp status is NOTSOLVED after conflictAnalyzeLP().