added writing variables in the right order for CIP-files
Previously fixed variables in CIP appeared in their natural
order. This makes it impossible to parse them, if a variables
depends on another variable that has not appeared earlier.
We now write the variables in topological sorting order.
We simply loop over the variables and write out the variables
that only depend on variables already written. This is not
the fastest way (because it has to loop over the variables
several times in the worst case), but it should be fast for
almost any instances, because the depth is usually very
small (say 0 or 1).