try to fix some memory limit problems
In order to allow for a clean exit of a (sub) SCIP if the memory
limit has been reached, we reserve at least 1% of the memory that
has already been used for the estimate of the memory that
is used by external components. Then SCIPisStopped() will
slightly earlier say that one should stop solving, leaving enough
room to finish the current node (without running into a hard
memory limit). The hope is that one gets a least a statistic output.