correct asserts in cons.c
In probing conflict constraints might be added that are initial, but
they are possibly valid only at a different node, i.e, they have to
be kept. In this case the number of nodes is still equal to the
original number, i.e., the previous asserts failed. The first assert
is fixed by checking whether we are in probing. In the second,
we check whether the number of initial constraints is 0.