replace old lint settings by new ones
Formerly, the files in lint/posix and lint/cpp contained definitions for
all important c functions. This is problematic if new functions are needed
and would need to be updated regularly. Lint provides a means to create
settings that are adapted to the particular compiler. This has been used
to create the new settings. See the lint documentation for more
information.

The new settings provides definitions for the gnu compiler, but in
principle can parse all common public include files. It thus avoids the
specific redefinitions of functions. However, one may possibly need to
add paths to the include files in the file gcc-include-path.lnt.