fixed hash functions of cons_linear and cutpool
in cons_linear parallel constraints where not detected as the hashfunction used the sign of the coefficients.
In the cutpool the hashfunction sometimes changed due to changed values of numminval and nummaxval. This
is hopefully fixed but still could occur if the maxval or minval change by too much. If this happens a
segfault can occur if the removal of a cut from the hashtable does not happen due to the changed hash. Then
the comparison function of the hashtable might get called for this already freed cut when a collision occurs later.
The proper solution would be to make sure that the minval maxval numminval nummaxval are all computed
deterministically and do not ever change later. Another solution would be to not use these values at all
in the hash function and the comparison functions of the rows in the cutpool.