changed asserts in lpFlushChgCols to deal with double infinity
if an LP solver (e.g., glop) uses infinities like
std::numeric_limits<T>::infinity() we cannot compute the
(relative) difference to another value, e.g., when using
SCIPis(Feas)EQ(). We need to check first if both values are
infinity w.r.t. to the LPI. If not, both values have to
be different to the internal infinity and equal.