add test for equal types in BMSduplicateXYZArray() functions in C++
In the BMSduplicateXYZArray() functions it can happen by accident
that the types of the array to be copied and the one to be copied
into are different. This can be checked at compile time with the
added checks.
The test only works for C++. For C, one would need to compile the
code in C++ mode. There are some workarounds for C, but they
all seem to fail if one argument is constant.